+++ src/main/java/Engine/Entity/NPC.java
@@ -65,7 +65,7 @@ public class NPC extends Entity {
             default -> null;
         };
 
-        graphics2D.drawImage(image, x, y, GameConstants.TILE_SIZE, GameConstants.TILE_SIZE, null);
+        graphics2D.drawImage(image, (int) x, (int) y, GameConstants.TILE_SIZE, GameConstants.TILE_SIZE, null);
         entitiesCollisionBox.drawCollider(graphics2D, Color.YELLOW);
         entitiesAggroZone.drawCollider(graphics2D, Color.BLACK);
     }
